Move provisioning to sub-package
