+++ jupyter_client/provisioning/__init__.py
@@ -1,0 +1,3 @@
+from .factory import KernelProvisionerFactory  # noqa
+from .local_provisioner import LocalProvisioner  # noqa
+from .provisioner_base import KernelProvisionerBase  # noqa
